requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
groq==0.4.2
twilio==8.12.0
python-dotenv==1.0.1
duckduckgo-search==4.4.3
python-multipart==0.0.9

sqlalchemy==2.0.25
sentence-transformers==2.3.1
numpy==1.26.4
